+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 - **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **express-session** has been upgraded from 1.17.0 to **1.17.2**
 - **grunt-contrib-compass** has been removed
+- **grunt-contrib-concat** has been removed
+- **grunt-contrib-uglify** has been removed
 - **grunt-contrib-watch** has been removed
 - **grunt-eslint** has been removed
 - **grunt-karma** has been removed